+++ kame/kame/rtadvd/rtadvd.c
@@ -266,7 +266,7 @@ main(argc, argv)
 		       "<%s> set timer to %ld:%ld. waiting for inputs " 
 		       "or timeout",
 		       __FUNCTION__,
-		       timeout->tv_sec, timeout->tv_usec);
+		       (long int)timeout->tv_sec, (long int)timeout->tv_usec);
 
 		if ((i = select(maxfd + 1, &select_fd,
 				NULL, NULL, timeout)) < 0) {
@@ -1403,7 +1403,8 @@ ra_timer_update(void *data, struct timeval *tm)
 
 	syslog(LOG_DEBUG,
 	       "<%s> RA timer on %s is set to %ld:%ld",
-	       __FUNCTION__, rai->ifname, tm->tv_sec, tm->tv_usec);
+	       __FUNCTION__, rai->ifname,
+	       (long int)tm->tv_sec, (long int)tm->tv_usec);
 
 	return;
 }
